#189: Revamp Generation structure
